Reorganize vector store module structure
