+++ packages/giselle-engine/src/core/vector-store/types/interface.ts
