+++ packages/giselle-engine/src/core/vector-store/types/object.ts
